requirements.txt
numpy~=1.26.3
pandas~=2.2.0
scikit-learn~=1.4.0
jupyter
sqlalchemy==2.0.25
passlib==1.7.4
bcrypt
python-multipart==0.0.6
boto3==1.34.19
fastapi~=0.109.0
pydantic~=2.5.3
python-dotenv~=1.0.0
boto3
botocore~=1.34.23